feat: implement client-side admin password protection
This commit introduces a client-side password protection mechanism for the admin page. A hardcoded password is used to grant access to the admin tools.

A prominent warning has been added to the code to highlight the insecurity of this method, as the password is exposed in the client-side JavaScript bundle.